// src/app/modules/admin/models/ship-from-locations.model.ts
export class ShipFromLocations {
  clientId: number = 0;
  locationId: number = 0;
  orgId: string = '';
  invOrgId: string = '';
  locationName: string = '';
  companyName: string = '';
  address1: string = '';
  address2: string = '';
  address3: string = '';
  address4: string = '';
  city: string = '';
  stateProvinceCode: string = '';
  postalCode: string = '';
  countryCode: string = '';
  contactName: string = '';
  phoneNumber: string = '';
  shipperName: string = '';
  fdxeAcctNo: string = '';
  fdxgAcctNo: string = '';
  upsAcctNo: string = '';
  dhlAcctNo: string = '';
  emailId: string = '';
}